Patch 2578992 by Andrew Church: "This patch fixes several bugs related to integer sizes on 64-bit systems (e.g. printf("%lX",u32) won't work correctly). It also removes an unused variable in src/gtk/gtkglwidget.c which I apparently missed in my previous fixes." Most of the changes are related to printf/scanf functions and it seems to work fine for me. Please double check it though :)
git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@1963 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/cheat.c
@@ -78,8 +78,8 @@ int CheatAddCode(int type, u32 addr, u32 val)
 
 int CheatAddARCode(const char *code)
 {
-   u32 addr;
-   u16 val;
+   unsigned long addr;
+   unsigned short val;
    sscanf(code, "%08lX %04hX", &addr, &val);
    switch (addr >> 28)
    {
@@ -180,8 +180,8 @@ int CheatRemoveCodeByIndex(int i)
 
 int CheatRemoveARCode(const char *code)
 {
-   u32 addr;
-   u16 val;
+   unsigned long addr;
+   unsigned short val;
    sscanf(code, "%08lX %04hX", &addr, &val);
 
    switch (addr >> 28)
