+++ yabause/src/debug.c
@@ -131,7 +131,7 @@ void DebugPrintf(Debug * d, const char * file, u32 line, const char * format, ..
   case DEBUG_STREAM:
     if (d->output.stream == NULL)
       break;
-    fprintf(d->output.stream, "%s (%s:%ld): ", d->name, file, line);
+    fprintf(d->output.stream, "%s (%s:%ld): ", d->name, file, (long)line);
     vfprintf(d->output.stream, format, l);
     break;
   case DEBUG_STRING:
@@ -140,7 +140,7 @@ void DebugPrintf(Debug * d, const char * file, u32 line, const char * format, ..
       if (d->output.string == NULL)
 	break;
       
-      i = sprintf(d->output.string, "%s (%s:%ld): ", d->name, file, line);
+      i = sprintf(d->output.string, "%s (%s:%ld): ", d->name, file, (long)line);
       vsprintf(d->output.string + i, format, l);
     }
     break;
@@ -148,7 +148,7 @@ void DebugPrintf(Debug * d, const char * file, u32 line, const char * format, ..
     {
       int i;
       int strnewhash = 0;
-      i = sprintf(strtmp, "%s (%s:%ld): ", d->name, file, line);
+      i = sprintf(strtmp, "%s (%s:%ld): ", d->name, file, (long)line);
       i += vsprintf(strtmp + i, format, l);
       for ( ; i>0 ; i-- ) strnewhash += (int)(strtmp[i]);
       if ( strnewhash != strhash ) d->output.callback( strtmp );
